% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmcirkelkort.R
\name{fmcirkelkort}
\alias{fmcirkelkort}
\title{Title}
\usage{
fmcirkelkort(
  data,
  lat,
  lon,
  label = NULL,
  radius = 3,
  color = NULL,
  scale = "factor",
  bins = 7
)
}
\arguments{
\item{data}{Choose the data frame}

\item{lat}{The name of the column in the data with the latitude coordinates.}

\item{lon}{The name of the column in the data with the longitude coordinates.}

\item{label}{Adds a lable to the to the center of the circle. Should be specified as a name of one of the columns of the data.}

\item{radius}{The radius of the circles on the map (in meters). If a numerical value is chosen, all circles will have the same radius. If a string is chosen, the column in the data with the name of the string will be chosen. The radius of the circles will be equal to the number in the chosen column. The column must contain numeric values. Defalut is 30000 meters.}

\item{color}{A string equal to a column name in the data. If specified, the circles will have different colors.}

\item{scale}{The type of the scale for the colors (if color is specified). Can be either 'factor', 'numeric' or 'bin'}

\item{bins}{The number of bins if 'bin' is specified in scale.}
}
\value{
something
}
\description{
Title
}
\examples{
\dontrun{
fmcirkelkort()
}
}
