% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmisokort.R
\name{fmisokort}
\alias{fmisokort}
\title{fmisokort}
\usage{
fmisokort(
  data = NULL,
  lat = NULL,
  lon = NULL,
  color = NULL,
  quota = 500,
  profile = "driving-car",
  range = 10,
  intervals = 1,
  api_key = NULL,
  legend = FALSE,
  output = NULL,
  map = NULL
)
}
\arguments{
\item{data}{Choose the data frame}

\item{lat}{The name of the column in the data with the latitude coordinates.}

\item{lon}{The name of the column in the data with the longitude coordinates.}

\item{color}{The color of the isochrones}

\item{quota}{Set the quota left on the api key}

\item{profile}{Choose the profile. This is the same as the profile for openrouteservice::ors_isochrones. Default = 'driving-car'.}

\item{range}{Choose the range in km for the isochrones. Defalut = 10.}

\item{intervals}{Choose the number of isochrones for each coorinate. Default = 1}

\item{api_key}{Set the api key}

\item{legend}{Add a legend? default = FALSE}

\item{output}{The name of the output (a .png file). If not specified, not outcome will b exported.}

\item{map}{Add the isochrones to a map.}
}
\value{
a map
}
\description{
fmisokort
}
\examples{
\dontrun{
fmisokort()
}
}
