% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmkommunekort.R
\name{fmkommunekort}
\alias{fmkommunekort}
\title{fmkommunekort}
\usage{
fmkommunekort(
  data = NULL,
  id = NULL,
  value = NULL,
  scale = "numeric",
  bins = 5,
  legend = FALSE,
  legendtitle = NULL,
  bornholm = T,
  output = NULL
)
}
\arguments{
\item{data}{The data frame for the municipality data}

\item{id}{The name of the municipality}

\item{value}{The value for the municipality which shall correspond to the color on the map}

\item{scale}{The type of the scale for the colors. Can be either 'factor', 'numeric' (default), 'bin.num' or 'bin.cat'}

\item{bins}{The number of bins if 'bin' specified under 'scale'. Default = 5.}

\item{legend}{Choose whether the map should includ a legend (default = FALSE)}

\item{legendtitle}{Choose the name above the legendtitle. Default is the same name as the column specified under 'legend'}

\item{bornholm}{Should Bornholm be moved north west to be included in the map?}

\item{output}{The name of the exported .png file. No map exported if the output = NULL (default = NULL)}
}
\value{
A map of the Danish municipalities.
}
\description{
fmkommunekort
}
\examples{
\dontrun{
fmkommunekort(data = data, id = id, value = value)
}
}
