% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrarkort.R
\name{fmrarkort}
\alias{fmrarkort}
\title{Title}
\usage{
fmrarkort(
  data = NULL,
  id = NULL,
  value = NULL,
  suffix = "",
  bornholm = TRUE,
  marker = TRUE,
  ndigits = 1
)
}
\arguments{
\item{data}{data}

\item{id}{id}

\item{value}{value}

\item{suffix}{suffix}

\item{bornholm}{bornholm}

\item{marker}{marker}

\item{ndigits}{digits}
}
\value{
a map
}
\description{
Title
}
\examples{
\dontrun{
fmrarkort(data = data, id = id, value = value)
}
}
