% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmregionkort.R
\name{fmregionkort}
\alias{fmregionkort}
\title{fmregionkort}
\usage{
fmregionkort(
  data = NULL,
  id = NULL,
  value = NULL,
  scale = "numeric",
  bins = 5,
  legend = FALSE,
  legendtitle = NULL,
  bornholm = T,
  marker = T,
  suffix = "",
  output = NULL,
  ndigits = 1
)
}
\arguments{
\item{data}{The data frame for the region data}

\item{id}{The name of the region}

\item{value}{The value for the region which shall correspond to the color on the map}

\item{scale}{The type of the scale for the colors. Can be either 'factor', 'numeric' (default), 'bin.num' or 'bin.cat'}

\item{bins}{The number of bins if 'bin' specified under 'scale'. Default = 5.}

\item{legend}{Choose whether the map should includ a legend (default = FALSE)}

\item{legendtitle}{Choose the name above the legendtitle. Default is the same name as the column specified under 'legend'}

\item{bornholm}{Should Bornholm be moved north west to be included in the map?}

\item{marker}{Adds marker to the map saying e.g. "Nordjylland: xx" where the value of xx correspond to the value for Nordjylland in the data. Defalut = TRUE}

\item{suffix}{The suffix of the marker text, e.g. "pct."}

\item{output}{The name of the exported .png file. No map exported if the output = NULL (default = NULL)}

\item{ndigits}{The number of digits on the marker text}
}
\value{
A map of the Danish regions
}
\description{
fmregionkort
}
\examples{
\dontrun{
fmregionkort(data = data, id = id, value = value)
}
}
