% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional}
\alias{regional}
\title{Danish municipalities}
\format{
A Large SpatialPolygonsDataFrame
}
\usage{
regional
}
\description{
A set of spatial polygons to form regions
}
\keyword{datasets}
