% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmisokort.R
\name{fmisokort}
\alias{fmisokort}
\title{Title}
\usage{
fmisokort(
  data = NULL,
  lat = NULL,
  lon = NULL,
  quota = 500,
  profile = "driving-car",
  range = 10,
  intervals = 1,
  range_type = "distance",
  api_key = NULL,
  output = NULL
)
}
\arguments{
\item{data}{data}

\item{lat}{lat}

\item{lon}{lon}

\item{quota}{quata}

\item{profile}{profile}

\item{range}{range}

\item{intervals}{interval}

\item{range_type}{range_type}

\item{api_key}{api}

\item{output}{output}
}
\value{
a map
}
\description{
Title
}
\examples{
\dontrun{
fmisokort()
}
}
