% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmkommunekort.R
\name{fmkommunekort}
\alias{fmkommunekort}
\title{Title}
\usage{
fmkommunekort(
  data = NULL,
  id = NULL,
  value = NULL,
  num_val = "numeric",
  bins = 5,
  legend = FALSE,
  legendtitle = NULL,
  output = NULL
)
}
\arguments{
\item{data}{x}

\item{id}{x}

\item{value}{x}

\item{num_val}{x}

\item{bins}{x}

\item{legend}{x}

\item{legendtitle}{x}

\item{output}{x}
}
\value{
A map
}
\description{
Title
}
\examples{
\dontrun{
fmkommunekort()
}
}
